wasm/{wasm_vector, plonk_proof?: move WasmVecVecF{p,q}
Move into wasm_vector instead of plonk_proof to reuse the structure in other
files, without depending on the whole macros
